+++ src/main/java/com/dabomstew/pkrandom/constants/GlobalConstants.java
@@ -93,6 +93,8 @@ public class GlobalConstants {
 
     public static final int[] ptSpecialIntros = { 377, 378, 379, 479, 480, 482, 483, 484, 485, 486, 487, 491, 492, 493 };
 
+    public static final List<Integer> SUPPORTED_GENERATIONS = Arrays.asList(1, 2, 3, 4, 5);
+
     public static final List<Integer> battleTrappingAbilities = Arrays.asList(23, 42, 71);
 
     public static final List<Integer> negativeAbilities = Arrays.asList(129, 112, 54, 59, 161, 150, 96, 100, 121, 103);